Feature added to allow the user to create a plan which consists of the Plan Name, Goal, Purpose.
Made the decision that objectives and tasks will be a separate feature and
is not mandatory for making a Plan. So I deleted the Objective and Task
classes for now.

Also made Plan and Goal implement Parcelable so I can pass them through
an intent to send back to the scrolling activity.